+++ integration-tests/src/test/java/org/realityforge/arez/integration/BasicRepositoryIntegrationTest.java
@@ -54,8 +54,8 @@ public class BasicRepositoryIntegrationTest
   {
     final ArezContext context = Arez.context();
 
-    final BasicRepositoryIntegrationTest$TestComponentRepository repository =
-      BasicRepositoryIntegrationTest$TestComponentRepository.newRepository();
+    final BasicRepositoryIntegrationTest_TestComponentRepository repository =
+      BasicRepositoryIntegrationTest_TestComponentRepository.newRepository();
     final TestComponent component1 = repository.create( 1, "S1" );
     final TestComponent component2 = repository.create( 2, "S2" );
 
@@ -126,8 +126,8 @@ public class BasicRepositoryIntegrationTest
   {
     final ArezContext context = Arez.context();
 
-    final BasicRepositoryIntegrationTest$TestComponentRepository repository =
-      BasicRepositoryIntegrationTest$TestComponentRepository.newRepository();
+    final BasicRepositoryIntegrationTest_TestComponentRepository repository =
+      BasicRepositoryIntegrationTest_TestComponentRepository.newRepository();
 
     // component1 has value that will sort after component2 to test sorting below
     final TestComponent component1 = repository.create( 1, "B" );
@@ -158,8 +158,8 @@ public class BasicRepositoryIntegrationTest
   {
     final ArezContext context = Arez.context();
 
-    final BasicRepositoryIntegrationTest$TestComponentRepository repository =
-      BasicRepositoryIntegrationTest$TestComponentRepository.newRepository();
+    final BasicRepositoryIntegrationTest_TestComponentRepository repository =
+      BasicRepositoryIntegrationTest_TestComponentRepository.newRepository();
 
     // component1 has value that will sort after component2 to test sorting below
     final TestComponent component1 = repository.create( 1, "B" );