+++ processor/src/test/resources/expected/com/example/repository/NestedModel_Arez_BasicActionModel.java
@@ -1,3 +1,5 @@
+package com.example.repository;
+
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import org.realityforge.arez.Arez;
@@ -7,23 +9,25 @@ import org.realityforge.arez.Disposable;
 import org.realityforge.braincheck.Guards;
 
 @Generated("org.realityforge.arez.processor.ArezProcessor")
-public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicActionModel implements Disposable {
+public final class NestedModel_Arez_BasicActionModel extends NestedModel.BasicActionModel implements Disposable {
   private static volatile long $$arez$$_nextId;
 
   private final long $$arez$$_id;
 
   private boolean $$arez$$_disposed;
 
+  private OnDispose $$arez$$_onDispose;
+
   @Nonnull
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;
 
-  public NestedModel$Arez_BasicActionModel() {
+  public NestedModel_Arez_BasicActionModel() {
     super();
     this.$$arez$$_context = Arez.context();
     this.$$arez$$_id = $$arez$$_nextId++;
-    this.$$arez$$_component = Arez.areNativeComponentsEnabled() ? this.$$arez$$_context.createComponent( "BasicActionModel", $$arez$$_id(), $$arez$$_name(), null, null ) : null;
+    this.$$arez$$_component = Arez.areNativeComponentsEnabled() ? this.$$arez$$_context.createComponent( "BasicActionModel", $$arez$$_id(), $$arez$$_name(), () -> $$arez$$_preDispose(), null ) : null;
     if ( Arez.areNativeComponentsEnabled() ) {
       this.$$arez$$_component.complete();
     }
@@ -50,9 +54,21 @@ public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicAc
         this.$$arez$$_component.dispose();
       } else {
         this.$$arez$$_context.safeAction( Arez.areNamesEnabled() ? $$arez$$_name() + ".dispose" : null, () -> { {
+          $$arez$$_preDispose();
         } } );
       }
     }
+  }
+
+  void $$arez$$_preDispose() {
+    if ( null != this.$$arez$$_onDispose ) {
+      this.$$arez$$_onDispose.onDispose( this );
+      this.$$arez$$_onDispose = null;
+    }
+  }
+
+  void $$arez$$_setOnDispose(OnDispose onDispose) {
+    this.$$arez$$_onDispose = onDispose;
   }
 
   @Override
@@ -80,10 +96,10 @@ public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicAc
   public final boolean equals(final Object o) {
     if ( this == o ) {
       return true;
-    } else if ( null == o || !(o instanceof NestedModel$Arez_BasicActionModel) ) {
+    } else if ( null == o || !(o instanceof NestedModel_Arez_BasicActionModel) ) {
       return false;
     } else {
-      final NestedModel$Arez_BasicActionModel that = (NestedModel$Arez_BasicActionModel) o;;
+      final NestedModel_Arez_BasicActionModel that = (NestedModel_Arez_BasicActionModel) o;;
       return $$arez$$_id() == that.$$arez$$_id();
     }
   }
@@ -95,5 +111,10 @@ public final class NestedModel$Arez_BasicActionModel extends NestedModel.BasicAc
     } else {
       return super.toString();
     }
+  }
+
+  @FunctionalInterface
+  interface OnDispose {
+    void onDispose(NestedModel_Arez_BasicActionModel entity);
   }
 }